Adds ObjectStoreProvider and dependent types.
I'm constraining my work to just OpenStack cloud files support here.

This is breaking ground for cloud files support.  We start by creating
an ObjectStoreProvider interface, which is empty at the moment since all
I'm doing is discovering the appropriate endpoint to use.  We create our
bindings in our context structure, and at the package scope as well.
Finally, we provide a simple stub of an integration test for it all.