Return a Container when creating a container.
The previous definition of ObjectStoreProvider followed OpenStack
conventions where the API was flat, and you passed specific objects into
relatively fixed endpoints.  However, this leaked the abstraction I
ultimately intend to deliver, which is to abstract away API-specific
details.  By refactoring the code so that Container structures are their
own things, and expose their own interfaces, we hide more of the
implementation details, and that should allow us greater freedom when it
comes time to implement support for, e.g., Google AppEngine or Amazon
S3.